requirements.txt
asgiref==3.2.10
boto3==1.14.13
botocore==1.17.13
Django==3.1.1
django-filter==2.3.0
django-storages==1.9.1
docutils==0.15.2
gunicorn==20.0.4
jmespath==0.10.0
Pillow==7.1.2
python-dateutil==2.8.1
pytz==2020.1
s3transfer==0.3.3
six==1.15.0
# sqlparse==0.3.1
psycopg2-binary==2.8.6
psycopg2==2.8.6
urllib3==1.25.9
whitenoise==5.1.0
django-widget-tweaks==1.4.8
